Modified to support general R-Factor and MOS.
git-svn-id: https://svn.code.sf.net/p/osp-toolkit/code/branches/3.5@496 43700839-f8e2-0310-9689-8ed4146226e7

+++ include/osp/osptransapi.h
@@ -109,8 +109,8 @@ extern "C" {
     int OSPPTransactionSetPackLossMean(OSPTTRANHANDLE, OSPE_DIRECTION, unsigned);
     int OSPPTransactionSetSrcNetworkId(OSPTTRANHANDLE, const char *);
     int OSPPTransactionSetDestNetworkId(OSPTTRANHANDLE, const char *);
-    int OSPPTransactionSetRFactor(OSPTTRANHANDLE, OSPE_DIRECTION, unsigned);
-    int OSPPTransactionSetMOS(OSPTTRANHANDLE, OSPE_DIRECTION, unsigned);
+    int OSPPTransactionSetRFactor(OSPTTRANHANDLE, OSPE_DIRECTION, float);
+    int OSPPTransactionSetMOS(OSPTTRANHANDLE, OSPE_DIRECTION, float);
 
 #ifdef __cplusplus
 }
